New Mage-Flow image model, diffusion editing (FLUX.2 / Krea-2 / Qwen-Image-Edit), and a no-toolchain Metal build
Highlights since the previous snapshot:

- Image generation & editing: a new microsoft/Mage-Flow family (an
  NR-MMDiT transformer + MageVAE + conditioner, wired for text-to-image
  and image editing, with an on-by-default provenance watermark). The
  FLUX.2, Krea-2, and Qwen-Image-Edit paths matured a lot: bf16
  transformers that match their references, image-editing flows
  (reference-image conditioning, an identity-edit LoRA, a Qwen3-VL
  deepstack), high-resolution VAE tiling fixes so 1024/2K fit in 16 GB,
  fused transpose+rope / layernorm+adaLN, and matrix-core (M5)
  acceleration for the diffusion GEMMs and attention.

- Gemma-4 12B: an OptiQ per-tensor mixed-precision bring-up spanning
  text, multimodal, and batched decode, with the M5 matmul2d path.

- Build: Metal kernels can now be compiled at runtime, so vpipe builds
  and runs without the Xcode Metal Toolchain. When metal/metallib are
  absent the build embeds the shader source and compiles each kernel on
  first use via the OS runtime compiler; force either mode with
  -DVPIPE_METAL_RUNTIME_COMPILE=ON|OFF.

- Pipelines & web UI: a model-select stage and a shared model port for
  composing conditioner/DiT/VAE; self-contained LoRA-fused models; an
  anti-aliased resample (lanczos) for VAE-encode downscales; a multi-line
  text-prompt config type; and a visual composer that saves/loads
  pipelines through the server, still-image preview with pan/zoom, and
  tab-scoped access keys.

- Robustness: the test runner now propagates a failure as the process
  exit status and contains an exception thrown out of a test; the model
  loader tolerates an index-listed shard being absent from disk.

c0f76ca2

+++ gpu-kernels/metal/gemm/dense_gemm.metal
@@ -49,10 +49,13 @@ kernel void dense_gemm_bias_f16(
   for (int k0 = 0; k0 < Kd; k0 += DG_TILE) {
     // As[ty][tx] = x[m, k0+tx]; Bs[ty][tx] = W[n_of_tx, k0+ty].
     const int ax = k0 + tx;
-    As[ty][tx] = (m < M && ax < Kd) ? x[(uint)m * Kd + ax] : (VPIPE_ELT)0;
+    // 64-bit row offsets: for a VAE-decode conv-GEMM the im2col operand x[M,Kd]
+    // has M*Kd > 2^32 once the output area passes ~1.86M px at Kd=9*256 (M =
+    // Hout*Wout, e.g. 2048x1536 -> 3.15M*2304 = 7.26G), which a uint index wraps.
+    As[ty][tx] = (m < M && ax < Kd) ? x[(ulong)m * Kd + ax] : (VPIPE_ELT)0;
     const int bn = (int)tgig.x * DG_TILE + ty;   // the W-row this slot feeds
     const int bk = k0 + tx;
-    Bs[ty][tx] = (bn < N && bk < Kd) ? W[(uint)bn * Kd + bk] : (VPIPE_ELT)0;
+    Bs[ty][tx] = (bn < N && bk < Kd) ? W[(ulong)bn * Kd + bk] : (VPIPE_ELT)0;
     threadgroup_barrier(mem_flags::mem_threadgroup);
     for (int t = 0; t < DG_TILE; ++t) {
       // x[m, k0+t] = As[ty][t]; W[n, k0+t] = Bs[tx][t].
@@ -62,7 +65,7 @@ kernel void dense_gemm_bias_f16(
   }
   if (m < M && n < N) {
     if (has_bias) { acc += (float)bias[n]; }
-    y[(uint)m * N + n] = (VPIPE_ELT)acc;
+    y[(ulong)m * N + n] = (VPIPE_ELT)acc;
   }
 }
